analytics-service/requirements.txt
flask
pymongo
mysql-connector-python
numpy
python-dotenv